fix: improve Python robustness — error handling, async safety, precision
- Replace 7 bare except:pass with logger.warning() in wqs.py, analyzer.py
- Fix 6 database connection leaks with try/finally: conn.close() patterns
- Fix Optional falsy coalescing (ROI=0 treated as None) in wqs.py, validator.py
- Bound _metrics_cache and _trades_cache with LRU eviction (maxlen=500)
- Add clear_all_caches() call after wallet processing loop
- Fix TOCTOU race in lock file cleanup — move stale check inside flock
- Align WalletRecord financial fields to Decimal (avg_trade_size_sol, etc.)
- Change LiquidityData float fields to Decimal (liquidity_usd, price_usd, volume_24h_usd)
- Convert auto_merge.py to async — asyncio.sleep, asyncio.to_thread for HTTP
- Add task reference tracking and shutdown cancellation in laserstream_client
- Add isinstance assertion guards on dexscreener API response parsing
- Pre-existing: ML pipeline changes, regime models cleanup

+++ scout/core/models.py
@@ -34,12 +34,11 @@ class TraderArchetype(Enum):
 class LiquidityData:
     """Snapshot of token liquidity at a point in time."""
     token_address: str
-    liquidity_usd: float
-    price_usd: float
-    volume_24h_usd: float
+    liquidity_usd: Decimal
+    price_usd: Decimal
+    volume_24h_usd: Decimal
     timestamp: datetime
     source: str = "unknown"
-    # New: Token creation time for sniper checks
     token_creation_timestamp: Optional[datetime] = None
